app/requirements.txt
mysqlclient
request
flask